+++ Sentry-master/Model/CoreData/Item+CoreDataProperties.swift
@@ -18,7 +18,7 @@ extension Item {
     }
 
     @NSManaged public var name: String?
-    @NSManaged public var id: Int64
+    @NSManaged public var id: Int
     @NSManaged public var user: User?
 
 }
